Basic Frame work implemented and basic text area class
Created a basic program frame work. The textArea class will be "contain" the text within. Idea is to create "block" representing leds or segments in a display out of a set amount of pixels, the text area should be created to be a direct dimention of these "blocks" i.e. 10 blocks high.