Block addition complete
Added the functions to fill the text area with blocks, currently this is done using alot of hard coded values. The text area is meant to be white, but now if 250 blocks are requested it will fill with red (the current colour of the blocks). There is no spacing between blocks yet or padding around the edges.