Changed Method adding blocks
The method noew takes into account the size of the text area and creates a number of blocks that will fit in with the value of the x or y divided by 25. For example if the text area X is 600 / 25 = 24.0. Meaning the method will make 24 blocks per row, it is the same for the other measurement. The addition of 5 extra px to the size of the block is to account for the padding, although there is a remainder as well so it does not look right.